Move more functionality in base image of devcontainer
This change removes the use of devcontainer features and switches to
a smaller base image.  We were using
the universal devcontainer image, which is gigantic. The 'base' image
is only about 300mb. Once we add docker-in-docker and the Azure CLI it
goes back up to around 2gb, but this is still smaller than where we were
before.

This is unfortunately as far as I could go to move functionality into
the base image. It's not possible to run docker-in-docker as part of the
devcontainer build, so everything that requires kubernetes needs to be
part of the creation command.