Add redis support to todo application
This change adds redis support to the tutorial application so that it
now supports in-memory, redis, and mongo.

There are NO changes to any tutorial content in this PR - this **only**
adds the functionality to use the todo application with redis. The
purpose of this is for more flexibility for demos. For example, we
should not make the tutorial confusing by talking about switching
between databases (not a common scenario).

The application determines which data store to use by using the name of
the connection. For example the tutorial uses the connection name
`itemstore`. Different databases can be configured via `itemstore`,
`redis`, or `mongo` as the connection name, setting the following
environment variables:

- `CONNECTION_ITEMSTORE_CONNECTIONSTRING` (current tutorial scenario)
- `CONNECTION_MONGODB_CONNECTIONSTRING` (mongo)
- `CONNECTION_REDIS_HOST` and others (redis)