Add new 'first application' demo app (#81)
This change includes a brand new demo application for Radius that we
intend to use to introduce Radius to user (run your first application).

This has two pages:

- Container info (demonstrates Radius features like env-vars)
- Todo (used to demonstrate recipes)

We also do logging which we'll use as part of the tutorial walkthrough.

This app is a rewrite of the existing todo application to use
create-react-app with a node backend. The reason for the rewrite is that
the existing todo sample is using a very very strange combo (Express +
EJS + Vue.js + Parcel) of javascript techs that are mostly obsolete
(except for Vue.js). Rather than wrestle with how to modernize the
patchwork of different techs, it makes more sense to adopt something
common (Express + React + Webpack) and well-maintained (create-react-app).