Clean up of sample connection code (#164)
This change removes dead/obsolete code from the tutorial app.

- Calling the connection 'itemstore' is from the previous version of the
  tutorial. Now we call it 'redis' or 'mongo'.
- Removing the **errorneous** use of the Redis connection string.

Redis connection strings are specific to the StackExchange Redis (.NET)
library, and not supported by the SDK we're using in TypeScript/Node.
The library we're using supports URLs which are a totally different
format that connection strings. We have an issue already tracking the
addition of URL to Radius.